chore: re-organize web server code
Reduces some use of global variables as well
